// sourcecode/BTrak.Packages/snova-payroll-widgets/module/module-components/src/lib/payrollwidgetsandreports/models/PayRollEmployeeResignationModel.ts
export class PayRollEmployeeResignationModel{
    employeeResignationId:string;
    resignationStatusId: string;
    employeeId:string;
    resignationDate:any;
    resignationApprovedById:string
    lastDate:any;
    approvedDate:any;
    isArchived: boolean;
    timeStamp: any;
    IsApproved: boolean;
    isWaitingForApproval: boolean;
    isRejected: boolean;
    profileImage: string;
    userId: string;
    employeeFullName:string;
    employeeNumber:string;
}